+++ docs/changelog.rst
@@ -19,6 +19,9 @@ Changelog
 - macOS: Allow scrolling window contents using mouse wheel/trackpad even when the
   window is not the active window (:iss:`729`)
 
+- Remote control: Allow matching windows by the environment variables of their
+  child process as well
+
 0.11.3 [2018-07-10]
 ------------------------------